+++ frontend/src/Author/Index/Menus/AuthorIndexSortMenu.js
@@ -20,15 +20,6 @@ function AuthorIndexSortMenu(props) {
     >
       <MenuContent>
         <SortMenuItem
-          name="status"
-          sortKey={sortKey}
-          sortDirection={sortDirection}
-          onPress={onSortSelect}
-        >
-          Monitored/Status
-        </SortMenuItem>
-
-        <SortMenuItem
           name="sortName"
           sortKey={sortKey}
           sortDirection={sortDirection}