+++ frontend/src/Author/Index/Posters/AuthorIndexPosters.js
@@ -37,7 +37,6 @@ function calculateRowHeight(posterHeight, sortKey, isSmallScreen, posterOptions)
   const {
     detailedProgressBar,
     showTitle,
-    showMonitored,
     showQualityProfile
   } = posterOptions;
 
@@ -51,10 +50,6 @@ function calculateRowHeight(posterHeight, sortKey, isSmallScreen, posterOptions)
   ];
 
   if (showTitle !== 'no') {
-    heights.push(19);
-  }
-
-  if (showMonitored) {
     heights.push(19);
   }
 
@@ -217,7 +212,6 @@ class AuthorIndexPosters extends Component {
     const {
       detailedProgressBar,
       showTitle,
-      showMonitored,
       showQualityProfile
     } = posterOptions;
 
@@ -244,7 +238,6 @@ class AuthorIndexPosters extends Component {
           posterHeight={posterHeight}
           detailedProgressBar={detailedProgressBar}
           showTitle={showTitle}
-          showMonitored={showMonitored}
           showQualityProfile={showQualityProfile}
           showRelativeDates={showRelativeDates}
           shortDateFormat={shortDateFormat}